buildfund_webapp/requirements.txt
django
djangorestframework
django-cors-headers
openai
requests
djongo>=1.3.6
pymongo>=3.12
python-dotenv